Refactored contacts on app level
