+++ src/App.jsx
@@ -3,6 +3,7 @@ import {
   BrowserRouter as Router,
   useRoutes,
 } from 'react-router-dom';
+import { LinkedIn, GitHub, Email } from '@mui/icons-material';
 import ErrorView from './views/ErrorView';
 import About from './views/About';
 import Skills_Tools from './views/SkillsTools';
@@ -12,11 +13,29 @@ import VRAR from './views/VRAR';
 import Header from './components/Header';
 import Footer from './components/Footer';
 
+const CONTACTS = [
+  {
+    contactMethod: 'LinkedIn',
+    icon: <LinkedIn />,
+    hrefLink: 'https://www.linkedin.com/in/thomas-a-mendez',
+  },
+  { 
+    contactMethod: 'GitHub',
+    icon: <GitHub />,
+    hrefLink: 'https://github.com/thomasmendez',
+  },
+  {
+    contactMethod: 'Email',
+    icon: <Email />,
+    hrefLink: 'mailto:thomasmendez01@gmail.com',
+  },
+];
+
 const AppRoutes = () => {
   const routes = useRoutes([
-    { path: '/', exact: true, element: <About /> },
-    { path: '/home', exact: true, element: <About /> },
-    { path: '/about', exact: true, element: <About /> },
+    { path: '/', exact: true, element: <About contacts={CONTACTS}/> },
+    { path: '/home', exact: true, element: <About contacts={CONTACTS}/> },
+    { path: '/about', exact: true, element: <About contacts={CONTACTS}/> },
     { path: '/skillsTools', exact: true, element: <Skills_Tools /> },
     { path: '/work', exact: true, element: <Work /> },
     { path: '/softwareEngineering', exact: true, element: <SoftwareEngineering /> },
@@ -31,7 +50,7 @@ const App = () => {
     <Router>
       <Header />
       <AppRoutes />
-      <Footer />
+      <Footer contacts={CONTACTS}/>
     </Router>
   );
 };
