+++ src/views/VRAR.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { Grid } from '@mui/material';
 import TitleHeader from '../components/TitleHeader';
 import Project from '../components/Projects/Project';
